Extract upload function into BALArchiver and separate metadata

+++ modules/wikidata.py
@@ -191,14 +191,14 @@ class BALMWikidata(object):
 
         Returns dict with all the item metadata or False if an error occurred.
         """
-        sitename = self.conv.getNameFromDB(database, pretext=True)
-        datename = self.conv.getDateFromWiki(dumpdate)
-        arcdate = self.conv.getDateFromWiki(dumpdate, archivedate=True)
         try:
             datetime.datetime.strptime(dumpdate, '%Y%m%d')
         except ValueError:
             self.common.giveMessage('The date was given in the wrong format!')
             return False
+        sitename = self.conv.getNameFromDB(database, pretext=True)
+        datename = self.conv.getDateFromWiki(dumpdate)
+        arcdate = self.conv.getDateFromWiki(dumpdate, archivedate=True)
 
         metadata = {
             'collection': self.config.get('collection'),
@@ -528,10 +528,13 @@ class BALMWikidata(object):
 
         Returns: True if process is successful, False if otherwise.
         """
-        count = 0
         identifier = "wikibase-%s-%s" % (database, dumpdate)
         iaitem = balchivist.BALArchiver(identifier)
         allfiles = self.getFiles(database, dumpdate)
+        md = self.getItemMetadata(database, dumpdate)
+        headers = {
+            'x-archive-size-hint': self.sizehint
+        }
 
         if self.resume:
             items = []
@@ -563,39 +566,12 @@ class BALMWikidata(object):
             return False
 
         os.chdir(dumps)
-        for dumpfile in items:
-            self.common.giveMessage("Uploading file: %s" % (dumpfile))
-            time.sleep(1)  # For Ctrl+C
-            if count == 0:
-                metadata = self.getItemMetadata(database, dumpdate)
-                headers = {
-                    'x-archive-size-hint': self.sizehint
-                }
-                upload = iaitem.uploadFile(dumpfile, metadata=metadata,
-                                           headers=headers)
-                # Allow the Internet Archive to process the item creation
-                if self.debug:
-                    pass
-                else:
-                    timenow = time.strftime("%Y-%m-%d %H:%M:%S",
-                                            time.localtime())
-                    self.common.giveMessage("Sleeping for 30 seconds, %s" %
-                                            (timenow))
-                    time.sleep(30)
-            else:
-                upload = iaitem.uploadFile(dumpfile)
+        upload = iaitem.upload(body=items, metadata=md, headers=headers)
 
-            if upload:
-                count += 1
-            else:
-                return False
-
-        if (path is None):
+        if (upload and path is None):
             shutil.rmtree(dumps)
         else:
-            pass
-
-        return True
+            return upload
 
     def check(self, database, dumpdate):
         """
